new docker tour

+++ lib/PerlChinaWebSite.pm
@@ -6,7 +6,7 @@ use List::MoreUtils qw(natatime);
 use Web::Query;
 use Data::Dumper;
 use File::Temp qw(tempfile);
-use IPC::Run qw(run timeout);
+use IPC::Run qw(start harness timeout);
 use Encode qw(decode encode);
 
 our $VERSION = '0.1';
@@ -22,12 +22,23 @@ get '/' => sub {
 ajax '/run' => sub {
     my ($in, $out, $err);
     my $code = param('code');
-    my @cmd = qw(docker run -v /tmp/:/tmp:ro ubuntu:perl-tour perl);
+    my @cmd = qw(docker run -m 128m -v /tmp/:/tmp:ro -u www pcws /run.sh);
     my ($fh, $temp) = tempfile();
     binmode($fh, ':utf8');
     print $fh $code;
+    chmod '0644', $temp;
     push @cmd, $temp;
-    run \@cmd, \$in, \$out, \$err, timeout(10) or debug($?);
+    my $h;
+    eval {
+        $h = harness \@cmd, \$in, \$out, \$err, ( my $t = timeout 10 );
+        start $h;
+        $h->finish;
+    };
+    if ($@) {
+        my $x = $@;
+        $h->kill_kill;
+        return $x;
+    };
     unlink $temp;
     return to_json({
         Errors => [ split(/\n/, decode('utf8', $err)) ],
